cogに対応
